+++ TaikoLocalServer/Services/Interfaces/ICardService.cs
@@ -5,15 +5,15 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface ICardService
 {
-    public Task<Card?> GetCardByAccessCode(string accessCode);
+	public Task<Card?> GetCardByAccessCode(string accessCode);
 
-    public uint GetNextBaid();
+	public ulong GetNextBaid();
 
-    public Task<List<User>> GetUsersFromCards();
+	public Task<List<User>> GetUsersFromCards();
 
-    public Task AddCard(Card card);
+	public Task AddCard(Card card);
 
-    public Task<bool> DeleteCard(string accessCode);
-
-    public Task<bool> UpdatePassword(string accessCode, string password, string salt);
+	public Task<bool> DeleteCard(string accessCode);
+	
+	public Task<bool> UpdatePassword(string accessCode, string password, string salt);
 }
